pylitterbot2mqtt/requirements.txt
paho-mqtt<2.0.0
asyncio
json2
pylitterbot
